scripts/requirements.txt
# Dependencias para los scripts de Alejardín

# Script generador de flores (generate_flowers.py)
# No requiere dependencias externas - solo usa stdlib

# Script de descarga de imágenes (download_flower_images.py)
requests==2.31.0        # HTTP requests a Pexels API
python-dotenv==1.0.1    # Cargar variables de entorno
Pillow==10.2.0          # Procesamiento y optimización de imágenes

# Dependencias opcionales para web scraping futuro:
# beautifulsoup4==4.12.3
# playwright==1.42.0
# lxml==5.1.0
